# col-cache: optimize for performance