<!-- Thanks for submitting a pull request! Please provide enough information so that others can review your pull request. The two fields below are mandatory. -->

## Summary

<!-- Explain the **motivation** for making this change. What existing problem does the pull request solve? -->
Doing a profiling in chrome dev tools shows that the function `decodeAddress` and `validateAddress` in `lib/utils/col-cache.js` are taking considerable amount of time, mainly on the cache lookup and regexp matching. This would make that faster, especally with large worksheets.

## Test plan

<!-- Demonstrate the code is solid. Example: The exact commands you ran and their output, screenshots / videos if the pull request changes UI. -->
Existing unit test and `npm run benchmark` result below.

Env: Windows 10 x64, node.js 12.18.4 LTS x64
<details> 
  <summary>npm run benchmark on master(a68209129a05b9ae37ca01ac705f84af76453088)</summary>


```
> exceljs@4.1.1 benchmark D:\UserData\projects\exceljs
> node --expose-gc benchmark


####################################################
WARMUP: Current memory usage: 8.89 MB
WARMUP: huge xlsx file streams profiling started
Reading worksheet 1
Reading row 50000
Reading row 100000
Reading worksheet 2
Reading row 150000
Processed 2 worksheets and 150002 rows
WARMUP: huge xlsx file streams profiling finished in 8354ms
WARMUP: Current memory usage (before GC): 153.16 MB
WARMUP: Current memory usage (after GC): 41.63 MB

####################################################
RUN 1: huge xlsx file streams profiling started
Reading worksheet 1
Reading row 50000
Reading row 100000
Reading worksheet 2
Reading row 150000
Processed 2 worksheets and 150002 rows
RUN 1: huge xlsx file streams profiling finished in 8174ms
RUN 1: Current memory usage (before GC): 116.57 MB
RUN 1: Current memory usage (after GC): 25.67 MB

####################################################
RUN 2: huge xlsx file streams profiling started
Reading worksheet 1
Reading row 50000
Reading row 100000
Reading worksheet 2
Reading row 150000
Processed 2 worksheets and 150002 rows
RUN 2: huge xlsx file streams profiling finished in 8048ms
RUN 2: Current memory usage (before GC): 142.09 MB
RUN 2: Current memory usage (after GC): 41.6 MB

####################################################
RUN 3: huge xlsx file streams profiling started
Reading worksheet 1
Reading row 50000
Reading row 100000
Reading worksheet 2
Reading row 150000
Processed 2 worksheets and 150002 rows
RUN 3: huge xlsx file streams profiling finished in 8242ms
RUN 3: Current memory usage (before GC): 115.43 MB
RUN 3: Current memory usage (after GC): 25.67 MB

####################################################
WARMUP: Current memory usage: 25.33 MB
WARMUP: huge xlsx file async iteration profiling started
Reading worksheet 1
Reading row 50000
Reading row 100000
Reading worksheet 2
Reading row 150000
Processed 2 worksheets and 150002 rows
WARMUP: huge xlsx file async iteration profiling finished in 8120ms
WARMUP: Current memory usage (before GC): 148.28 MB
WARMUP: Current memory usage (after GC): 41.65 MB

####################################################
RUN 1: huge xlsx file async iteration profiling started
Reading worksheet 1
Reading row 50000
Reading row 100000
Reading worksheet 2
Reading row 150000
Processed 2 worksheets and 150002 rows
RUN 1: huge xlsx file async iteration profiling finished in 8181ms
RUN 1: Current memory usage (before GC): 110.38 MB
RUN 1: Current memory usage (after GC): 25.67 MB

####################################################
RUN 2: huge xlsx file async iteration profiling started
Reading worksheet 1
Reading row 50000
Reading row 100000
Reading worksheet 2
Reading row 150000
Processed 2 worksheets and 150002 rows
RUN 2: huge xlsx file async iteration profiling finished in 8215ms
RUN 2: Current memory usage (before GC): 148.74 MB
RUN 2: Current memory usage (after GC): 41.64 MB

####################################################
RUN 3: huge xlsx file async iteration profiling started
Reading worksheet 1
Reading row 50000
Reading row 100000
Reading worksheet 2
Reading row 150000
Processed 2 worksheets and 150002 rows
RUN 3: huge xlsx file async iteration profiling finished in 8259ms
RUN 3: Current memory usage (before GC): 112.38 MB
RUN 3: Current memory usage (after GC): 25.68 MB


```

</details>


<details> 
  <summary>npm run benchmark on myfreeer:col-cache-performance(8a47ae7a7e32cc709998e8d7bbe478dccaa0be62)</summary>


```

> exceljs@4.1.1 benchmark D:\UserData\projects\exceljs
> node --expose-gc benchmark


####################################################
WARMUP: Current memory usage: 8.89 MB
WARMUP: huge xlsx file streams profiling started
Reading worksheet 1
Reading row 50000
Reading row 100000
Reading worksheet 2
Reading row 150000
Processed 2 worksheets and 150002 rows
WARMUP: huge xlsx file streams profiling finished in 6030ms
WARMUP: Current memory usage (before GC): 67.6 MB
WARMUP: Current memory usage (after GC): 9.65 MB

####################################################
RUN 1: huge xlsx file streams profiling started
Reading worksheet 1
Reading row 50000
Reading row 100000
Reading worksheet 2
Reading row 150000
Processed 2 worksheets and 150002 rows
RUN 1: huge xlsx file streams profiling finished in 5787ms
RUN 1: Current memory usage (before GC): 53.34 MB
RUN 1: Current memory usage (after GC): 9.75 MB

####################################################
RUN 2: huge xlsx file streams profiling started
Reading worksheet 1
Reading row 50000
Reading row 100000
Reading worksheet 2
Reading row 150000
Processed 2 worksheets and 150002 rows
RUN 2: huge xlsx file streams profiling finished in 5926ms
RUN 2: Current memory usage (before GC): 41.67 MB
RUN 2: Current memory usage (after GC): 9.76 MB

####################################################
RUN 3: huge xlsx file streams profiling started
Reading worksheet 1
Reading row 50000
Reading row 100000
Reading worksheet 2
Reading row 150000
Processed 2 worksheets and 150002 rows
RUN 3: huge xlsx file streams profiling finished in 5808ms
RUN 3: Current memory usage (before GC): 40.32 MB
RUN 3: Current memory usage (after GC): 9.76 MB

####################################################
WARMUP: Current memory usage: 9.47 MB
WARMUP: huge xlsx file async iteration profiling started
Reading worksheet 1
Reading row 50000
Reading row 100000
Reading worksheet 2
Reading row 150000
Processed 2 worksheets and 150002 rows
WARMUP: huge xlsx file async iteration profiling finished in 5808ms
WARMUP: Current memory usage (before GC): 48.61 MB
WARMUP: Current memory usage (after GC): 9.77 MB

####################################################
RUN 1: huge xlsx file async iteration profiling started
Reading worksheet 1
Reading row 50000
Reading row 100000
Reading worksheet 2
Reading row 150000
Processed 2 worksheets and 150002 rows
RUN 1: huge xlsx file async iteration profiling finished in 5837ms
RUN 1: Current memory usage (before GC): 38.44 MB
RUN 1: Current memory usage (after GC): 9.77 MB

####################################################
RUN 2: huge xlsx file async iteration profiling started
Reading worksheet 1
Reading row 50000
Reading row 100000
Reading worksheet 2
Reading row 150000
Processed 2 worksheets and 150002 rows
RUN 2: huge xlsx file async iteration profiling finished in 5800ms
RUN 2: Current memory usage (before GC): 51.21 MB
RUN 2: Current memory usage (after GC): 9.85 MB

####################################################
RUN 3: huge xlsx file async iteration profiling started
Reading worksheet 1
Reading row 50000
Reading row 100000
Reading worksheet 2
Reading row 150000
Processed 2 worksheets and 150002 rows
RUN 3: huge xlsx file async iteration profiling finished in 5793ms
RUN 3: Current memory usage (before GC): 48.64 MB
RUN 3: Current memory usage (after GC): 9.83 MB


```

</details>